BIC-187 # skip persistent steps if no good storage - avoids harmless errors popping up in the console

+++ src/bic/model-application.js
@@ -187,6 +187,10 @@ define(function (require) {
                 }
               ),
               function (model) {
+                if (!app.hasStorage()) {
+                  app.interactions.remove(model);
+                  return Promise.resolve();
+                }
                 return new Promise(function (resolve, reject) {
                   model.destroy({
                     success: resolve,
@@ -201,6 +205,9 @@ define(function (require) {
           function () {
             app.forms.whenUpdated();
             app.retrieveDataSuitcasesForInteractions();
+            if (!app.hasStorage()) {
+              return Promise.resolve();
+            }
             return app.interactions.save();
           }
         );
